Halt script execution when for non 200 responses
The fetch library we're using only throws an error if there are network
issues, but 4xx and 5xx responses are not network issues, so no error is
thrown.

To protect against changes in API behavior, unless the HTTP requests
explicitly return 2xx error codes, the script stops execution.

This change is made to the new DappRadar code as well as existing
Coingecko API calls.