Don't store by chain ID (or even CoinType)
We don't know which chain IDs or coinTypes will be added by OFAC
upstream, and it's difficult given an arbitrary symbol to identify which
chain ID / CoinType it is. So it's easier to just put all the banned
addresses in a flat namespace. There's only 540 so it's not really a
performance hit to check all of them for an Ethereum transaction instead
of just the ETH ones.